Pulled LanguageInfo struct and helper functions into RageUtil_LanguageInfo files
