+++ src/RageUtil.cpp
@@ -412,169 +412,6 @@ RString vssprintf( const char *szFormat, va_list argList )
 	return sRet;
 }
 
-/* ISO-639-1 codes: http://www.loc.gov/standards/iso639-2/php/code_list.php
- * We don't use 3-letter codes, so we don't bother supporting them. */
-static const LanguageInfo g_langs[] =
-{
-	{"aa", "Afar"},
-	{"ab", "Abkhazian"},
-	{"af", "Afrikaans"},
-	{"am", "Amharic"},
-	{"ar", "Arabic"},
-	{"as", "Assamese"},
-	{"ay", "Aymara"},
-	{"az", "Azerbaijani"},
-	{"ba", "Bashkir"},
-	{"be", "Byelorussian"},
-	{"bg", "Bulgarian"},
-	{"bh", "Bihari"},
-	{"bi", "Bislama"},
-	{"bn", "Bengali"},
-	{"bo", "Tibetan"},
-	{"br", "Breton"},
-	{"ca", "Catalan"},
-	{"co", "Corsican"},
-	{"cs", "Czech"},
-	{"cy", "Welsh"},
-	{"da", "Danish"},
-	{"de", "German"},
-	{"dz", "Bhutani"},
-	{"el", "Greek"},
-	{"en", "English"},
-	{"eo", "Esperanto"},
-	{"es", "Spanish"},
-	{"et", "Estonian"},
-	{"eu", "Basque"},
-	{"fa", "Persian"},
-	{"fi", "Finnish"},
-	{"fj", "Fiji"},
-	{"fo", "Faeroese"},
-	{"fr", "French"},
-	{"fy", "Frisian"},
-	{"ga", "Irish"},
-	{"gd", "Gaelic"},
-	{"gl", "Galician"},
-	{"gn", "Guarani"},
-	{"gu", "Gujarati"},
-	{"ha", "Hausa"},
-	{"he", "Hebrew"},
-	{"hi", "Hindi"},
-	{"hr", "Croatian"},
-	{"hu", "Hungarian"},
-	{"hy", "Armenian"},
-	{"ia", "Interlingua"},
-	{"id", "Indonesian"},
-	{"ie", "Interlingue"},
-	{"ik", "Inupiak"},
-	{"in", "Indonesian"}, // compatibility
-	{"is", "Icelandic"},
-	{"it", "Italian"},
-	{"iw", "Hebrew"}, // compatibility
-	{"ja", "Japanese"},
-	{"ji", "Yiddish"}, // compatibility
-	{"jw", "Javanese"},
-	{"ka", "Georgian"},
-	{"kk", "Kazakh"},
-	{"kl", "Greenlandic"},
-	{"km", "Cambodian"},
-	{"kn", "Kannada"},
-	{"ko", "Korean"},
-	{"ks", "Kashmiri"},
-	{"ku", "Kurdish"},
-	{"ky", "Kirghiz"},
-	{"la", "Latin"},
-	{"ln", "Lingala"},
-	{"lo", "Laothian"},
-	{"lt", "Lithuanian"},
-	{"lv", "Latvian"},
-	{"mg", "Malagasy"},
-	{"mi", "Maori"},
-	{"mk", "Macedonian"},
-	{"ml", "Malayalam"},
-	{"mn", "Mongolian"},
-	{"mo", "Moldavian"},
-	{"mr", "Marathi"},
-	{"ms", "Malay"},
-	{"mt", "Maltese"},
-	{"my", "Burmese"},
-	{"na", "Nauru"},
-	{"ne", "Nepali"},
-	{"nl", "Dutch"},
-	{"no", "Norwegian"},
-	{"oc", "Occitan"},
-	{"om", "Oromo"},
-	{"or", "Oriya"},
-	{"pa", "Punjabi"},
-	{"pl", "Polish"},
-	{"ps", "Pashto"},
-	{"pt", "Portuguese"},
-	{"qu", "Quechua"},
-	{"rm", "Rhaeto-Romance"},
-	{"rn", "Kirundi"},
-	{"ro", "Romanian"},
-	{"ru", "Russian"},
-	{"rw", "Kinyarwanda"},
-	{"sa", "Sanskrit"},
-	{"sd", "Sindhi"},
-	{"sg", "Sangro"},
-	{"sh", "Serbo-Croatian"},
-	{"si", "Singhalese"},
-	{"sk", "Slovak"},
-	{"sl", "Slovenian"},
-	{"sm", "Samoan"},
-	{"sn", "Shona"},
-	{"so", "Somali"},
-	{"sq", "Albanian"},
-	{"sr", "Serbian"},
-	{"ss", "Siswati"},
-	{"st", "Sesotho"},
-	{"su", "Sudanese"},
-	{"sv", "Swedish"},
-	{"sw", "Swahili"},
-	{"ta", "Tamil"},
-	{"te", "Tegulu"},
-	{"tg", "Tajik"},
-	{"th", "Thai"},
-	{"ti", "Tigrinya"},
-	{"tk", "Turkmen"},
-	{"tl", "Tagalog"},
-	{"tn", "Setswana"},
-	{"to", "Tonga"},
-	{"tr", "Turkish"},
-	{"ts", "Tsonga"},
-	{"tt", "Tatar"},
-	{"tw", "Twi"},
-	{"uk", "Ukrainian"},
-	{"ur", "Urdu"},
-	{"uz", "Uzbek"},
-	{"vi", "Vietnamese"},
-	{"vo", "Volapuk"},
-	{"wo", "Wolof"},
-	{"xh", "Xhosa"},
-	{"yi", "Yiddish"},
-	{"yo", "Yoruba"},
-	{"zh-Hans", "Chinese (Simplified)"},
-	{"zh-Hant", "Chinese (Traditional)"},
-	{"zu", "Zulu"},
-};
-
-void GetLanguageInfos( std::vector<const LanguageInfo*> &vAddTo )
-{
-	for( unsigned i=0; i<ARRAYLEN(g_langs); ++i )
-		vAddTo.push_back( &g_langs[i] );
-}
-
-const LanguageInfo *GetLanguageInfo( const RString &sIsoCode )
-{
-	for( unsigned i=0; i<ARRAYLEN(g_langs); ++i )
-	{
-		if( EqualsNoCase(sIsoCode, g_langs[i].szIsoCode) )
-			return &g_langs[i];
-	}
-
-	return nullptr;
-}
-
 RString join( const RString &sDeliminator, const std::vector<RString> &sSource)
 {
 	if( sSource.empty() )
