+++ src/ScreenOptionsMasterPrefs.cpp
@@ -17,6 +17,7 @@
 #include "LocalizedString.h"
 #include "SpecialFiles.h"
 #include "RageLog.h"
+#include "RageUtil_LanguageInfo.h"
 
 #include <vector>
